move app outside entry point file
